+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.0] - 2026-08-05
+
+### Added
+- New Home tab to display New Arrivals, Mixes, and Daily Discovery
+
+### Changed
+- Playlists now use v2 API for better support.
+- Pagination now fetches next page of tracks only once when approaching end of list.
+
 ## [0.10.0] - 2026-08-04
 
 ### Added